added Repository and fix entities

+++ src/main/java/com/beeauto/entities/Remarque.java
@@ -14,7 +14,7 @@ public class Remarque {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
 
-    private int idRemarque;
+    private long idRemarque;
     private String text;
     @Transient
     private Date dateCreation;
@@ -22,7 +22,7 @@ public class Remarque {
     public Remarque() {
     }
 
-    public Remarque(int idRemarque, String text) {
+    public Remarque(long idRemarque, String text) {
         this.idRemarque = idRemarque;
         this.text = text;
     }
@@ -31,11 +31,11 @@ public class Remarque {
         this.text = text;
     }
 
-    public int getIdRemarque() {
+    public long getIdRemarque() {
         return idRemarque;
     }
 
-    public void setIdRemarque(int idRemarque) {
+    public void setIdRemarque(long idRemarque) {
         this.idRemarque = idRemarque;
     }
 
@@ -54,6 +54,7 @@ public class Remarque {
     public void setDateCreation(Date dateCreation) {
         this.dateCreation = dateCreation;
     }
+
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "idDemandeAbonnement", nullable = false)
